Move apiFetchSaga actions to mockBackendSaga
- apiFetchSaga is removed
- deprecated some mock userAPI enums
