+++ src/mocks/backend.ts
@@ -1,10 +1,12 @@
 import { delay, SagaIterator } from 'redux-saga'
-import { put, takeEvery } from 'redux-saga/effects'
+import { call, put, select, takeEvery } from 'redux-saga/effects'
 
 import * as actions from '../actions'
 import * as actionTypes from '../actions/actionTypes'
+import { IState } from '../reducers/states'
 import { history } from '../utils/history'
 import { mockAssessmentOverviews, mockAssessments } from './assessmentAPI'
+import { mockFetchGrading, mockFetchGradingOverview } from './gradingAPI'
 
 export function* mockBackendSaga(): SagaIterator {
   yield takeEvery(actionTypes.FETCH_AUTH, function*(action) {
@@ -13,8 +15,8 @@ export function* mockBackendSaga(): SagaIterator {
       refreshToken: 'refreshToken'
     }
     const user = {
-      name: 'DevStudent',
-      role: 'student'
+      name: 'DevStaff',
+      role: 'staff'
     }
     yield put(actions.setTokens(tokens))
     yield put(actions.setRole(user.role))
@@ -31,5 +33,22 @@ export function* mockBackendSaga(): SagaIterator {
     const id = (action as actionTypes.IAction).payload
     const assessment = mockAssessments[id]
     yield put(actions.updateAssessment(assessment))
+  })
+
+  yield takeEvery(actionTypes.FETCH_GRADING_OVERVIEWS, function*() {
+    const accessToken = yield select((state: IState) => state.session.accessToken)
+    const gradingOverviews = yield call(() => mockFetchGradingOverview(accessToken))
+    if (gradingOverviews !== null) {
+      yield put(actions.updateGradingOverviews(gradingOverviews))
+    }
+  })
+
+  yield takeEvery(actionTypes.FETCH_GRADING, function*(action) {
+    const submissionId = (action as actionTypes.IAction).payload
+    const accessToken = yield select((state: IState) => state.session.accessToken)
+    const grading = yield call(() => mockFetchGrading(accessToken, submissionId))
+    if (grading !== null) {
+      yield put(actions.updateGrading(submissionId, grading))
+    }
   })
 }
